refactor(frontend): promote personalization/storage → lib/personalization/
Bean pawrrtal-lqs4. `features/personalization/storage.ts` (60 LOC of
persistence + option catalogs) and `hooks/use-personalization.ts` (the
React Query wrapper around the same endpoints) were imported by 9 files
across 3 feature modules (`features/onboarding/v2/*`,
`features/settings/sections/*`), behaving as shared lib code while
nominally living inside a feature. This PR moves both — plus their
co-located tests — down to `lib/personalization/`.

The 9 importer sites are rewritten to `@/lib/personalization/*`. No UI
moves: onboarding steps and the settings section stay where they are;
only the data layer relocates.

`frontend/features/personalization/` is removed entirely (no files left).

Score impact: small. Sentrux's modularity component is dominated by the
~750 cross-module import edges in the graph; this PR doesn't reduce the
edge count (the 9 callers still cross a module boundary, just to lib/
instead of features/personalization/) so quality_signal stays at 6912.
What it does fix: the high-fan-in node `features/personalization/storage`
is no longer concentrated inside one feature, which lifts the equality
distribution slightly. The architectural correctness — lib-level data
code in lib/ — is the durable win.

bun run typecheck + bun run arch:check: clean.
20 test files (83 tests) across lib/personalization/, features/onboarding/,
features/settings/: all green.
