+++ frontend/features/onboarding/v2/step-context.tsx
@@ -4,7 +4,7 @@ import { ArrowRight } from 'lucide-react';
 import type * as React from 'react';
 import { Button } from '@/components/ui/button';
 import { Textarea } from '@/components/ui/textarea';
-import type { PersonalizationProfile } from '@/features/personalization/storage';
+import type { PersonalizationProfile } from '@/lib/personalization/storage';
 import { OnboardingShell } from './onboarding-shell';
 
 /** Props for {@link StepContext}. */
