+++ frontend/features/onboarding/v2/step-identity.tsx
@@ -3,7 +3,7 @@
 import type * as React from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import type { PersonalizationProfile } from '@/features/personalization/storage';
+import type { PersonalizationProfile } from '@/lib/personalization/storage';
 import { cn } from '@/lib/utils';
 import { OnboardingShell } from './onboarding-shell';
