+++ frontend/features/onboarding/v2/step-messaging.tsx
@@ -5,12 +5,12 @@ import type * as React from 'react';
 import { useCallback, useEffect, useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { TelegramConnectDialog } from '@/features/channels/TelegramConnectDialog';
+import { listChannels } from '@/lib/channels';
 import {
 	MESSAGING_CHANNELS,
 	type MessagingChannelId,
 	type PersonalizationProfile,
-} from '@/features/personalization/storage';
-import { listChannels } from '@/lib/channels';
+} from '@/lib/personalization/storage';
 import { cn } from '@/lib/utils';
 import { OnboardingShell } from './onboarding-shell';
