+++ frontend/features/onboarding/v2/step-server.tsx
@@ -5,8 +5,8 @@ import type * as React from 'react';
 import { useCallback, useId, useReducer } from 'react';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import type { PersonalizationProfile } from '@/features/personalization/storage';
 import { API_BASE_URL, saveBackendConfig } from '@/lib/api';
+import type { PersonalizationProfile } from '@/lib/personalization/storage';
 import { cn } from '@/lib/utils';
 import { OnboardingShell } from './onboarding-shell';
